[Refactor] Split twitter_connector in others modules
